+++ src/main/java/org/synyx/urlaubsverwaltung/account/AccountRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 /**
  * Repository for {@link Account} entities.
  */
-public interface AccountRepository extends CrudRepository<AccountEntity, Integer> {
+public interface AccountRepository extends CrudRepository<AccountEntity, Long> {
 
     @Query("select x from account x where YEAR(x.validFrom) = ?1 and x.person = ?2")
     Optional<AccountEntity> getHolidaysAccountByYearAndPerson(int year, Person person);
