+++ src/main/java/org/synyx/urlaubsverwaltung/department/DepartmentService.java
@@ -24,7 +24,7 @@ public interface DepartmentService {
      * @param departmentId id of a {@link Department} to check
      * @return <code>true</code> if the departmentId exists, <code>false</code> otherwise
      */
-    boolean departmentExists(Integer departmentId);
+    boolean departmentExists(Long departmentId);
 
     /**
      * Returns a department by its unique identifier
@@ -32,7 +32,7 @@ public interface DepartmentService {
      * @param departmentId the unique identifier to of a department
      * @return department to given id
      */
-    Optional<Department> getDepartmentById(Integer departmentId);
+    Optional<Department> getDepartmentById(Long departmentId);
 
     /**
      * adds the given department to repository.
@@ -53,7 +53,7 @@ public interface DepartmentService {
      *
      * @param departmentId the unique identifier to delete a department
      */
-    void delete(Integer departmentId);
+    void delete(Long departmentId);
 
     /**
      * @return all departments ordered by the department name
@@ -172,7 +172,7 @@ public interface DepartmentService {
      * @param pageableSearchQuery searchQuery to restrict the result set
      * @return all managed and active members for the person
      */
-    Page<Person> getManagedMembersOfPersonAndDepartment(Person person, Integer departmentId, PageableSearchQuery pageableSearchQuery);
+    Page<Person> getManagedMembersOfPersonAndDepartment(Person person, Long departmentId, PageableSearchQuery pageableSearchQuery);
 
     /**
      * Check the role of the given person and return a {@link Page} of all managed and inactive {@link Person}s for the
@@ -184,7 +184,7 @@ public interface DepartmentService {
      * @param pageableSearchQuery search query containing pageable and an optional query for firstname/lastname
      * @return all managed and inactive members for the person
      */
-    Page<Person> getManagedInactiveMembersOfPersonAndDepartment(Person person, Integer departmentId, PageableSearchQuery pageableSearchQuery);
+    Page<Person> getManagedInactiveMembersOfPersonAndDepartment(Person person, Long departmentId, PageableSearchQuery pageableSearchQuery);
 
     /**
      * Get all distinct managed members of the department head.
