+++ src/main/java/org/synyx/urlaubsverwaltung/person/basedata/PersonBasedataEntity.java
@@ -18,7 +18,7 @@ class PersonBasedataEntity {
 
     @Id
     @Column(name = "person_id")
-    private Integer personId;
+    private Long personId;
 
     @OneToOne(fetch = LAZY)
     @PrimaryKeyJoinColumn(name = "person_id", referencedColumnName = "id")
@@ -33,11 +33,11 @@ class PersonBasedataEntity {
         // OK
     }
 
-    public Integer getPersonId() {
+    public Long getPersonId() {
         return personId;
     }
 
-    public void setPersonId(Integer personId) {
+    public void setPersonId(Long personId) {
         this.personId = personId;
     }
