+++ src/main/java/org/synyx/urlaubsverwaltung/person/basedata/PersonBasedataService.java
@@ -15,7 +15,7 @@ public interface PersonBasedataService {
      * @param personId to find base data of the person
      * @return base data of the given person
      */
-    Optional<PersonBasedata> getBasedataByPersonId(int personId);
+    Optional<PersonBasedata> getBasedataByPersonId(long personId);
 
     /**
      * Returns an optional person base data object if there is
@@ -24,7 +24,7 @@ public interface PersonBasedataService {
      * @param personIds to find base data of the given persons
      * @return base data map of the given persons grouped by the id of the person
      */
-    Map<PersonId, PersonBasedata> getBasedataByPersonId(List<Integer> personIds);
+    Map<PersonId, PersonBasedata> getBasedataByPersonId(List<Long> personIds);
 
     /**
      * Update a person base data object
